+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 * fix(BREAKING): [#1818](https://github.com/lambdaclass/cairo-vm/pull/1818): 
     * Fix `MemorySegmentManager::add_zero_segment` function when resizing a segment
-    * Signature change: `MemorySegmentManager::get_memory_holes` now receives  `builtin_segment_indexes: HashSet<usize>`
+    * Signature change(BREAKING): `MemorySegmentManager::get_memory_holes` now receives `builtin_segment_indexes: HashSet<usize>`
 
 * fix(BREAKING): Replace `CairoRunner` method `initialize_all_builtins` with `initialize_program_builtins`. Now it only initializes program builtins instead of all of them